+++ packages/cli/src/commands/block/publish.ts
@@ -56,6 +56,5 @@ export async function handler({
     }
 
     await publishBlock(config, ignoreConflict);
-    logger.info('Finished extracting messages.');
   }
 }